+++ config/ban.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of Laravel Ban.
+ *
+ * (c) Anton Komarev <anton@komarev.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Ban Database Migrations
+    |--------------------------------------------------------------------------
+    |
+    | Determine if default package migrations should be registered.
+    | Set value to `false` when using customized migrations.
+    |
+    */
+
+    'load_default_migrations' => true,
+
+];