+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to `cybercog/laravel-ban` will be documented in this file.
 
 ## [Unreleased]
 
+## [4.2.0] - 2019-09-26
+
+### Added
+
+- ([#50]) Added publishable configuration file
+- ([#50]) Added ability to ignore default migrations
+
 ## [4.1.0] - 2019-09-04
 
 ### Added
@@ -118,7 +125,8 @@ All notable changes to `cybercog/laravel-ban` will be documented in this file.
 
 - Initial release
 
-[Unreleased]: https://github.com/cybercog/laravel-ban/compare/4.1.0...master
+[Unreleased]: https://github.com/cybercog/laravel-ban/compare/4.2.0...master
+[4.2.0]: https://github.com/cybercog/laravel-ban/compare/4.1.0...4.2.0
 [4.1.0]: https://github.com/cybercog/laravel-ban/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/cybercog/laravel-ban/compare/3.5.0...4.0.0
 [3.5.0]: https://github.com/cybercog/laravel-ban/compare/3.4.0...3.5.0
@@ -131,6 +139,7 @@ All notable changes to `cybercog/laravel-ban` will be documented in this file.
 [2.0.1]: https://github.com/cybercog/laravel-ban/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/cybercog/laravel-ban/compare/1.0.0...2.0.0
 
+[#50]: https://github.com/cybercog/laravel-ban/pull/50
 [#48]: https://github.com/cybercog/laravel-ban/pull/48
 [#35]: https://github.com/cybercog/laravel-ban/pull/35
 [#30]: https://github.com/cybercog/laravel-ban/pull/30